requirements.txt
scikit-learn~=1.1.1
numpy~=1.22.4
pandas~=1.4.2
torch
tqdm~=4.64.0
ax-platform~=0.2.5.1
psutil~=5.9.1
seaborn~=0.11.2
ax~=0.52.0
scipy~=1.8.1
torchmetrics~=0.8.2
jupyter
pyyaml>=5.4
shap~=0.41.0
biopython~=1.79
captum
